MDL-56515 admin: hide settings links when not applicable
As settings can be defined using a settings.php file or using the config_form() function
we can not rely on the existence of a file into the plugin file tree.

By introducing an has_config() into auth_plugin_base like for blocks in block_base
we allow plugin developers to explicitly defined if their plugins has settings or not.

Contrary to blocks the default has been set to true to allow backward compatibility.