MDL-51546 enrol_ldap: pagesize warning on empty value
admin_setting_configtext_trim_lower() does not allow to specify the data type,
so when pagesize is set to empty it is considered as a "valid" value and does not correcly set the default value.

The warning is caused by ldap_control_paged_result() which expect pagesize to be an int but receive an other type.